pyproject.toml
[tool.poetry]
name = "diagnostyka-wyniki-crawler"
version = "0.1.0"
description = "Skeleton for the diagnostyka wyniki crawler."
authors = ["diagnostyka-wyniki team <info@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.12.11"
playwright = "^1.55.0"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4"

[tool.poetry.scripts]
wyniki-crawler = "diagnostyka_wyniki_crawler.crawler:main"

[build-system]
requires = ["poetry-core>=1.6.0"]
build-backend = "poetry.core.masonry.api"
